+++ tests/formatting/comment_spacing.out.vhd
@@ -1,0 +1,10 @@
+entity e is
+end entity e;   -- one
+
+architecture a of e is
+
+  signal x : bit;   -- two
+
+begin
+
+end architecture a;