+++ tests/formatting/comment_spacing.vhd
@@ -1,0 +1,6 @@
+entity e is
+end entity e; -- one
+architecture a of e is
+  signal x : bit; -- two
+begin
+end architecture a;